// types/types.ts
export interface SchoolInfo {
  // Basic Information
  name: string
  tagline: string
  motto: string
  
  // Contact Information
  address: string
  city: string
  phone: string
  email: string
  
  // Social Media (only those used in public website)
  facebook: string
  instagram: string
  twitter: string
  youtube: string
  
  // About Information
  description: string
  mission: string
  vision: string
  
  // Hero Section
  heroTitle: string
  heroSubtitle: string
  heroCTA: string
}